+++ Main.go
@@ -1,15 +1,15 @@
 package main
 
 import (
-	"NASP/DataStructures/BloomFilter"
-	"NASP/DataStructures/CacheLRU"
-	"NASP/DataStructures/CountMinSketch"
-	"NASP/DataStructures/HyperLogLog"
-	"NASP/DataStructures/Memtable"
-	"NASP/DataStructures/MerkleTree"
-	"NASP/DataStructures/SSTable"
-	"NASP/DataStructures/TokenBucket"
-	"NASP/DataStructures/WAL"
+	"NASP/Structures/BloomFilter"
+	"NASP/Structures/CacheLRU"
+	"NASP/Structures/CountMinSketch"
+	"NASP/Structures/HyperLogLog"
+	"NASP/Structures/Memtable"
+	"NASP/Structures/MerkleTree"
+	"NASP/Structures/SSTable"
+	"NASP/Structures/TokenBucket"
+	"NASP/Structures/WAL"
 	"bufio"
 	"encoding/json"
 	"fmt"
