+++ Structures/BloomFilter/BloomFilter.go
